pkg/*,internal/pkg/flags: move watch flags from pkg/internal/flags to internal/pkg/flags
