+++ pkg/ansible/flags/flag.go
@@ -17,14 +17,14 @@ package flags
 import (
 	"strings"
 
-	"github.com/operator-framework/operator-sdk/pkg/internal/flags"
+	"github.com/operator-framework/operator-sdk/internal/pkg/flags/watch"
 	"github.com/operator-framework/operator-sdk/pkg/log/zap"
 	"github.com/spf13/pflag"
 )
 
 // AnsibleOperatorFlags - Options to be used by an ansible operator
 type AnsibleOperatorFlags struct {
-	flags.WatchFlags
+	watch.WatchFlags
 	InjectOwnerRef bool
 	MaxWorkers     int
 }
